feat: refactor bunker scene and fluid page for modularity and enhanced visuals
- Introduced a new `bunker-world` module to centralize world geometry, node definitions, and pathfinding helpers.
- Created a `bunker-scene` module to encapsulate reusable components like `Ground`, `BoxMarker`, `AgentMesh`, and `Building`.
- Updated `BunkerPage` and `BunkerFluidPage` to utilize the new modules, improving code organization and maintainability.
- Enhanced visual elements in the bunker scene, including improved building configurations and animations for objects like keys and C4.
- Implemented pickup animations and explosion effects for a more dynamic user experience.

These changes streamline the codebase and enhance the visual fidelity of the bunker environment, improving overall interactivity.
