+++ examples/app/src/app/bunker/page.tsx
@@ -5,121 +5,10 @@ import * as THREE from 'three'
 import { Canvas, useFrame } from '@react-three/fiber'
 import { Agent, Task } from 'mahler'
 import { CameraControls } from '@react-three/drei'
-import { Geometry, Base, Subtraction, Addition } from '@react-three/csg'
+import { N, NodeId, BUILDINGS, NODE_POS, findPath, isImmediatellyReachable, Vec3 } from '../../lib/bunker-world'
+import { Ground, BoxMarker, Building, LabelSprite, AgentMesh, EnhancedObject, SmallSphere, InventoryItem, PickupAnimation } from '../../lib/bunker-scene'
 
 const AGENT_INSIDE_BUILDING_OPACITY = 0.5
-
-type Vec3 = [number, number, number]
-
-// Scenario nodes
-const N = Object.freeze({
-  COURTYARD: 'courtyard',
-  TABLE: 'table_area',
-  STORAGE_DOOR: 'storage_door',
-  STORAGE_INT: 'storage_interior',
-  C4_TABLE: 'c4_table',
-  BUNKER_DOOR: 'bunker_door',
-  BUNKER_INT: 'bunker_interior',
-  STAR: 'star_pos',
-  SAFE: 'safe_spot',
-} as const)
-
-type NodeId = (typeof N)[keyof typeof N]
-
-// Building configuration
-type BuildingConfig = {
-  center: Vec3
-  size: [number, number, number] // [width, height, depth]
-  doorFace: 'north' | 'south' | 'east' | 'west'
-  doorOffset?: number // distance from building edge for door positioning
-  doorSize: [number, number] // [width, height] of the door
-}
-
-const BUILDINGS: Record<string, BuildingConfig> = {
-  STORAGE: {
-    center: [-10, 0, 8],
-    size: [6, 3.5, 4.5],
-    doorFace: 'east',
-    doorOffset: 1.5, // meters from building edge
-    doorSize: [1.2, 1.6] // [width, height] - standard door
-  },
-  BUNKER: {
-    center: [15, 0, 0],
-    size: [7, 5, 7],
-    doorFace: 'west',
-    doorOffset: 1.5, // meters from building edge
-    doorSize: [1.8, 2.4] // [width, height] - larger bunker door
-  }
-}
-
-// Helper functions for building geometry calculations
-function getBuildingInteriorPosition(building: BuildingConfig, offsetFromCenter: Vec3 = [0, 0, 0]): Vec3 {
-  const [centerX, centerY, centerZ] = building.center
-  const [offsetX, offsetY, offsetZ] = offsetFromCenter
-  return [centerX + offsetX, centerY + offsetY, centerZ + offsetZ]
-}
-
-function getBuildingDoorPosition(building: BuildingConfig): Vec3 {
-  const [centerX, centerY, centerZ] = building.center
-  const [width, height, depth] = building.size
-  const offset = building.doorOffset || 0
-
-  switch (building.doorFace) {
-    case 'east':
-      return [centerX + width / 2 + offset, centerY, centerZ]
-    case 'west':
-      return [centerX - width / 2 - offset, centerY, centerZ]
-    case 'south':
-      return [centerX, centerY, centerZ + depth / 2 + offset]
-    case 'north':
-      return [centerX, centerY, centerZ - depth / 2 - offset]
-  }
-}
-
-// Additional helper to get building bounds for validation
-function getBuildingBounds(building: BuildingConfig): { min: Vec3; max: Vec3 } {
-  const [centerX, centerY, centerZ] = building.center
-  const [width, height, depth] = building.size
-  return {
-    min: [centerX - width / 2, centerY - height / 2, centerZ - depth / 2],
-    max: [centerX + width / 2, centerY + height / 2, centerZ + depth / 2]
-  }
-}
-
-// Helper to check if a position is inside a building
-function isPositionInsideBuilding(position: Vec3, building: BuildingConfig): boolean {
-  const bounds = getBuildingBounds(building)
-  const [x, y, z] = position
-  return (
-    x >= bounds.min[0] && x <= bounds.max[0] &&
-    y >= bounds.min[1] && y <= bounds.max[1] &&
-    z >= bounds.min[2] && z <= bounds.max[2]
-  )
-}
-
-// Calculated positions for each node
-const NODE_POS: Record<NodeId, Vec3> = {
-  // Fixed outdoor positions
-  [N.COURTYARD]: [0, 0, 0],
-  [N.TABLE]: [-10, 0, 0],
-//   [N.SAFE]: [0, 0, -10],
-  [N.SAFE]: (() => {
-    const pos = getBuildingDoorPosition(BUILDINGS.BUNKER)
-    return [pos[0] - 5, pos[1], pos[2]]
-  })(),
-
-  // Storage building positions (mathematically calculated)
-  // Storage building spans: [-12.5 to -7.5, 0, 6.25 to 9.75]
-  [N.STORAGE_DOOR]: getBuildingDoorPosition(BUILDINGS.STORAGE), // [-5.5, 0, 8] (east side + 2m offset)
-  [N.STORAGE_INT]: getBuildingInteriorPosition(BUILDINGS.STORAGE), // [-10, 0, 8] (center)
-  [N.C4_TABLE]: getBuildingInteriorPosition(BUILDINGS.STORAGE, [-1, 0, 0]), // [-12, 0, 8] (inside, left of center)
-
-  // Bunker building positions (mathematically calculated)
-  // Bunker building spans: [12.25 to 17.75, 0, -2 to 2]
-  [N.BUNKER_DOOR]: getBuildingDoorPosition(BUILDINGS.BUNKER), // [9.25, 0, 0] (west side - 3m offset)
-  [N.BUNKER_INT]: getBuildingInteriorPosition(BUILDINGS.BUNKER), // [15, 0, 0] (center)
-  [N.STAR]: getBuildingInteriorPosition(BUILDINGS.BUNKER, [2, 0, 0]), // [17, 0, 0] (inside, right of center)
-}
 
 // Adjacency and gates
 type WorldState = {
@@ -146,532 +35,6 @@ const initial: WorldState = {
   storageUnlocked: false,
   c4Placed: false,
   bunkerBreached: false,
-}
-
-type Edge = [NodeId, NodeId, (s: WorldState) => boolean]
-const RAW_EDGES: Edge[] = [
-  [N.COURTYARD, N.TABLE, () => true],
-  [N.COURTYARD, N.STORAGE_DOOR, () => true],
-  [N.COURTYARD, N.BUNKER_DOOR, () => true],
-  [N.COURTYARD, N.SAFE, () => true],
-  [N.TABLE, N.STORAGE_DOOR, () => true],
-  [N.STORAGE_DOOR, N.STORAGE_INT, (s) => s.storageUnlocked === true],
-  [N.STORAGE_INT, N.C4_TABLE, () => true],
-  [N.STORAGE_DOOR, N.BUNKER_DOOR, () => true],
-  [N.BUNKER_DOOR, N.BUNKER_INT, (s) => s.bunkerBreached === true],
-  [N.BUNKER_DOOR, N.SAFE, () => true],
-  [N.BUNKER_INT, N.STAR, () => true],
-]
-
-function makeAdjacency(raw: Edge[]) {
-  const map: Record<string, Array<{ to: NodeId; when: (s: WorldState) => boolean }>> = {}
-  for (const [a, b, when] of raw) {
-    ;(map[a] ||= []).push({ to: b, when })
-    ;(map[b] ||= []).push({ to: a, when })
-  }
-  return map
-}
-const ADJ = makeAdjacency(RAW_EDGES)
-
-function neighbors(state: WorldState, from: NodeId) {
-  return (ADJ[from] || [])
-    .filter((e) => e.when(state))
-    .map((e) => e.to)
-}
-
-function isImmediatellyReachable(state: WorldState, from: NodeId, to: NodeId) {
-  return neighbors(state, from).includes(to)
-}
-
-function findPath(state: WorldState, from: NodeId, to: NodeId): NodeId[] | null {
-  if (from === to) return [from]
-  const seen = new Set<NodeId>([from])
-  const q: NodeId[] = [from]
-  const prev = new Map<NodeId, NodeId>()
-  while (q.length) {
-    const cur = q.shift()!
-    for (const n of neighbors(state, cur)) {
-      if (seen.has(n)) continue
-      seen.add(n)
-      prev.set(n, cur)
-      if (n === to) {
-        const path = [to]
-        let p = prev.get(to)
-        while (p !== undefined) {
-          path.push(p)
-          p = prev.get(p)
-        }
-        path.reverse()
-        return path
-      }
-      q.push(n)
-    }
-  }
-  return null
-}
-
-// Simple ground plane - expanded for larger space
-function Ground() {
-  return (
-    <mesh rotation={[-Math.PI / 2, 0, 0]} receiveShadow>
-      <planeGeometry args={[60, 60]} />
-      <meshStandardMaterial color="#1f2937" />
-    </mesh>
-  )
-}
-
-function BoxMarker({ position, color = '#34495e', label }: { position: Vec3; color?: string; label: string }) {
-  return (
-    <group position={position}>
-      <mesh castShadow receiveShadow>
-        <boxGeometry args={[2, 0.4, 2]} />
-        <meshStandardMaterial color={color} />
-      </mesh>
-      {/*
-	  <mesh position={[0, 0.35, 0]}>
-        <boxGeometry args={[2, 0.02, 2]} />
-        <meshStandardMaterial color="#95a5a6" />
-      </mesh>
-	  */}
-      <LabelSprite position={[0, 0.9, 0]} text={label} />
-    </group>
-  )
-}
-
-function SmallSphere({ position, color = 'gold', visible = true, size = 0.18 }: { position: Vec3; color?: string; visible?: boolean; size?: number }) {
-  if (!visible) return null
-  return (
-    <mesh position={position} castShadow>
-      <sphereGeometry args={[size, 16, 16]} />
-      <meshStandardMaterial color={color} emissive={color} emissiveIntensity={0.3} />
-    </mesh>
-  )
-}
-
-function EnhancedObject({ position, color, type, visible = true, size = 0.25 }: {
-  position: Vec3;
-  color: string;
-  type: 'key' | 'c4' | 'star';
-  visible?: boolean;
-  size?: number;
-}) {
-  if (!visible) return null
-
-  return (
-    <group position={position}>
-      {/* Main sphere */}
-      <mesh castShadow>
-        <sphereGeometry args={[size, 20, 20]} />
-        <meshStandardMaterial
-          color={color}
-          emissive={color}
-          emissiveIntensity={0.4}
-          metalness={type === 'key' ? 0.8 : 0.2}
-          roughness={type === 'key' ? 0.2 : 0.4}
-        />
-      </mesh>
-
-      {/* Glowing aura */}
-      <mesh>
-        <sphereGeometry args={[size * 1.3, 12, 12]} />
-        <meshStandardMaterial
-          color={color}
-          emissive={color}
-          emissiveIntensity={0.1}
-          transparent
-          opacity={0.3}
-        />
-      </mesh>
-
-      {/* Floating animation */}
-      <AnimatedFloat yOffset={0.1} speed={2}>
-        <mesh position={[0, 0.05, 0]}>
-          <sphereGeometry args={[size * 0.7, 8, 8]} />
-          <meshStandardMaterial
-            color={color}
-            emissive={color}
-            emissiveIntensity={0.6}
-            transparent
-            opacity={0.6}
-          />
-        </mesh>
-      </AnimatedFloat>
-    </group>
-  )
-}
-
-function AnimatedFloat({ children, yOffset = 0.1, speed = 1 }: { children: React.ReactNode; yOffset?: number; speed?: number }) {
-  const ref = useRef<THREE.Group>(null!)
-
-  useFrame((state) => {
-    if (ref.current) {
-      ref.current.position.y = Math.sin(state.clock.elapsedTime * speed) * yOffset
-      ref.current.rotation.y = state.clock.elapsedTime * 0.5
-    }
-  })
-
-  return <group ref={ref}>{children}</group>
-}
-
-function InventoryItem({
-  agentPos,
-  type,
-  color,
-  index,
-  size = 0.15
-}: {
-  agentPos: Vec3;
-  type: 'key' | 'c4' | 'star';
-  color: string;
-  index: number;
-  size?: number;
-}) {
-  const ref = useRef<THREE.Group>(null!)
-
-  useFrame((state) => {
-    if (ref.current) {
-      // Position items in a circle above the agent
-      const angle = (index * Math.PI * 2) / 3 + state.clock.elapsedTime * 0.5 // Rotate slowly
-      const radius = 0.8
-      const height = 1.8 + Math.sin(state.clock.elapsedTime * 2 + index) * 0.1 // Bobbing motion
-
-      ref.current.position.set(
-        agentPos[0] + Math.cos(angle) * radius,
-        agentPos[1] + height,
-        agentPos[2] + Math.sin(angle) * radius
-      )
-    }
-  })
-
-  return (
-    <group ref={ref}>
-      <mesh castShadow>
-        <sphereGeometry args={[size, 16, 16]} />
-        <meshStandardMaterial
-          color={color}
-          emissive={color}
-          emissiveIntensity={0.5}
-          metalness={type === 'key' ? 0.8 : 0.2}
-          roughness={type === 'key' ? 0.2 : 0.4}
-        />
-      </mesh>
-
-      {/* Glowing trail effect */}
-      <mesh>
-        <sphereGeometry args={[size * 1.2, 8, 8]} />
-        <meshStandardMaterial
-          color={color}
-          emissive={color}
-          emissiveIntensity={0.2}
-          transparent
-          opacity={0.4}
-        />
-      </mesh>
-    </group>
-  )
-}
-
-function PickupAnimation({
-  animation,
-  onComplete
-}: {
-  animation: {
-    active: boolean
-    startPos: Vec3
-    endPos: Vec3
-    startTime: number
-    duration: number
-    type: 'key' | 'c4' | 'star'
-    color: string
-  };
-  onComplete: () => void;
-}) {
-  const ref = useRef<THREE.Group>(null!)
-
-  useFrame(() => {
-    if (!animation.active || !ref.current) return
-
-    const now = performance.now()
-    const elapsed = now - animation.startTime
-    const progress = Math.min(elapsed / animation.duration, 1)
-
-    // Smooth easing
-    const eased = 1 - Math.pow(1 - progress, 3)
-
-    // Interpolate position with arc
-    const start = new THREE.Vector3(...animation.startPos)
-    const end = new THREE.Vector3(...animation.endPos)
-    const mid = start.clone().lerp(end, 0.5)
-    mid.y += 1.5 // Arc height
-
-    let currentPos: THREE.Vector3
-    if (eased < 0.5) {
-      currentPos = start.clone().lerp(mid, eased * 2)
-    } else {
-      currentPos = mid.clone().lerp(end, (eased - 0.5) * 2)
-    }
-
-    ref.current.position.copy(currentPos)
-    ref.current.scale.setScalar(1 - eased * 0.3) // Shrink as it approaches agent
-
-    if (progress >= 1) {
-      onComplete()
-    }
-  })
-
-  if (!animation.active) return null
-
-  return (
-    <group ref={ref}>
-      <mesh castShadow>
-        <sphereGeometry args={[0.2, 12, 12]} />
-        <meshStandardMaterial
-          color={animation.color}
-          emissive={animation.color}
-          emissiveIntensity={0.6}
-        />
-      </mesh>
-    </group>
-  )
-}
-
-function LabelSprite({ position, text, color = '#ffffff', bg = 'rgba(0,0,0,0.55)' }: { position: Vec3; text: string; color?: string; bg?: string }) {
-  const textureRef = useRef<THREE.CanvasTexture | null>(null)
-  const canvasRef = useRef<HTMLCanvasElement | null>(null)
-
-  if (textureRef.current == null) {
-    const canvas = document.createElement('canvas')
-    canvas.width = 512
-    canvas.height = 128
-    canvasRef.current = canvas
-    const tex = new THREE.CanvasTexture(canvas)
-    tex.minFilter = THREE.LinearFilter
-    tex.magFilter = THREE.LinearFilter
-    tex.needsUpdate = true
-    textureRef.current = tex
-  }
-
-  useEffect(() => {
-    const canvas = canvasRef.current!
-    const ctx = canvas.getContext('2d')!
-    ctx.clearRect(0, 0, canvas.width, canvas.height)
-    // background
-    ctx.fillStyle = bg
-    const padX = 24
-    const padY = 16
-    ctx.fillRect(0, 0, canvas.width, canvas.height)
-    // text
-    ctx.fillStyle = color
-    ctx.font = 'bold 56px system-ui, -apple-system, Segoe UI, Roboto, sans-serif'
-    ctx.textAlign = 'center'
-    ctx.textBaseline = 'middle'
-    ctx.fillText(text, canvas.width / 2, canvas.height / 2 + 6)
-    textureRef.current!.needsUpdate = true
-  }, [text, color, bg])
-
-  useEffect(() => () => textureRef.current?.dispose(), [])
-
-  return (
-    <sprite position={position} scale={[2.4, 0.6, 1]}>
-      <spriteMaterial map={textureRef.current!} transparent depthWrite={false} />
-    </sprite>
-  )
-}
-
-type Face = 'north' | 'south' | 'east' | 'west'
-
-function Building({
-  center,
-  size,
-  color = '#4b5563',
-  label,
-  doorFace,
-  doorSize,
-  doorColor = '#a78bfa',
-  showDoor = true,
-  opacity = 1,
-  debug = false,
-}: {
-  center: Vec3
-  size: [number, number, number]
-  color?: string
-  label: string
-  doorFace: Face
-  doorSize: [number, number]
-  doorColor?: string
-  showDoor?: boolean
-  opacity?: number
-  debug?: boolean
-}) {
-  const [dx, dy, dz] = size
-  const wallThickness = 0.2
-  const doorThickness = 0.15
-
-  function getDoorGeometry(face: Face): { position: Vec3; rotation: [number, number, number] } {
-    const surfaceOffset = wallThickness / 2 + 0.01
-    const y = -dy / 2 + doorSize[1] / 2 // Door bottom aligns with building bottom
-
-    switch (face) {
-      case 'east':
-        return {
-          position: [dx / 2 + surfaceOffset, y, 0],
-          rotation: [0, Math.PI / 2, 0],
-        }
-      case 'west':
-        return {
-          position: [-dx / 2 - surfaceOffset, y, 0],
-          rotation: [0, Math.PI / 2, 0],
-        }
-      case 'south':
-        return {
-          position: [0, y, dz / 2 + surfaceOffset],
-          rotation: [0, 0, 0],
-        }
-      case 'north':
-      default:
-        return {
-          position: [0, y, -dz / 2 - surfaceOffset],
-          rotation: [0, 0, 0],
-        }
-    }
-  }
-
-  const doorGeom = getDoorGeometry(doorFace)
-  const isOpen = !showDoor
-
-  // Debug colors for each wall
-  const debugColors = {
-    floor: '#ff9500',    // orange
-    roof: '#8e44ad',     // purple
-    north: '#e74c3c',    // red
-    south: '#3498db',    // blue
-    east: '#2ecc71',     // green
-    west: '#f1c40f',     // yellow
-  }
-  const floorColor = debugColors.floor
-
-  return (
-    <group position={[center[0], dy / 2, center[2]]}>
-      {/* Floor */}
-      <mesh position={[0, -dy / 2, 0]} castShadow receiveShadow>
-        <boxGeometry args={[dx, wallThickness, dz]} />
-        <meshStandardMaterial color={debug ? debugColors.floor : floorColor} transparent opacity={opacity} />
-      </mesh>
-
-      {/* Roof */}
-      <mesh position={[0, dy / 2, 0]} castShadow receiveShadow>
-        <boxGeometry args={[dx, wallThickness, dz]} />
-        <meshStandardMaterial color={debug ? debugColors.roof : color} transparent opacity={opacity} />
-      </mesh>
-
-      {/* North Wall - RED - full width */}
-      <mesh castShadow receiveShadow position={[0, 0, -(dz / 2 - wallThickness / 2)]}>
-        <Geometry>
-          <Base>
-            <boxGeometry args={[dx, dy, wallThickness]} />
-          </Base>
-          {/* Cut door if this is the door wall and door is open */}
-          {doorFace === 'north' && isOpen && (
-            <Subtraction position={[0, -dy / 2 + doorSize[1] / 2, 0]}>
-              <boxGeometry args={[doorSize[0], doorSize[1], wallThickness + 0.1]} />
-            </Subtraction>
-          )}
-        </Geometry>
-        <meshStandardMaterial color={debug ? debugColors.north : color} transparent opacity={opacity} />
-      </mesh>
-
-      {/* South Wall - BLUE - full width */}
-      <mesh castShadow receiveShadow position={[0, 0, dz / 2 - wallThickness / 2]}>
-        <Geometry>
-          <Base>
-            <boxGeometry args={[dx, dy, wallThickness]} />
-          </Base>
-          {/* Cut door if this is the door wall and door is open */}
-          {doorFace === 'south' && isOpen && (
-            <Subtraction position={[0, -dy / 2 + doorSize[1] / 2, 0]}>
-              <boxGeometry args={[doorSize[0], doorSize[1], wallThickness + 0.1]} />
-            </Subtraction>
-          )}
-        </Geometry>
-        <meshStandardMaterial color={debug ? debugColors.south : color} transparent opacity={opacity} />
-      </mesh>
-
-      {/* East Wall - GREEN - shortened to fit between north/south walls */}
-      <mesh castShadow receiveShadow position={[dx / 2 - wallThickness / 2, 0, 0]}>
-        <Geometry>
-          <Base>
-            <boxGeometry args={[wallThickness, dy, dz - wallThickness]} />
-          </Base>
-          {/* Cut door if this is the door wall and door is open */}
-          {doorFace === 'east' && isOpen && (
-            <Subtraction position={[0, -dy / 2 + doorSize[1] / 2, 0]}>
-              <boxGeometry args={[wallThickness + 0.1, doorSize[1], doorSize[0]]} />
-            </Subtraction>
-          )}
-        </Geometry>
-        <meshStandardMaterial color={debug ? debugColors.east : color} transparent opacity={opacity} />
-      </mesh>
-
-      {/* West Wall - YELLOW - shortened to fit between north/south walls */}
-      <mesh castShadow receiveShadow position={[-(dx / 2 - wallThickness / 2), 0, 0]}>
-        <Geometry>
-          <Base>
-            <boxGeometry args={[wallThickness, dy, dz - wallThickness]} />
-          </Base>
-          {/* Cut door if this is the door wall and door is open */}
-          {doorFace === 'west' && isOpen && (
-            <Subtraction position={[0, -dy / 2 + doorSize[1] / 2, 0]}>
-              <boxGeometry args={[wallThickness + 0.1, doorSize[1], doorSize[0]]} />
-            </Subtraction>
-          )}
-        </Geometry>
-        <meshStandardMaterial color={debug ? debugColors.west : color} transparent opacity={opacity} />
-      </mesh>
-
-      {/* When door is closed, show contrasting door surface */}
-      {showDoor && (
-        <mesh position={doorGeom.position} rotation={doorGeom.rotation} castShadow>
-          <boxGeometry args={[doorSize[0], doorSize[1], doorThickness]} />
-          <meshStandardMaterial
-            color={doorColor}
-            metalness={0.3}
-            roughness={0.7}
-          />
-        </mesh>
-      )}
-
-      {/* Door frame for visual enhancement - only show when door is closed */}
-      {showDoor && (
-        <group position={doorGeom.position} rotation={doorGeom.rotation}>
-          {/* Door frame outline */}
-          <mesh castShadow>
-            <boxGeometry args={[doorSize[0] + 0.1, doorSize[1] + 0.1, doorThickness * 0.5]} />
-            <meshStandardMaterial
-              color="#2d3748"
-              metalness={0.5}
-              roughness={0.3}
-            />
-          </mesh>
-        </group>
-      )}
-
-      <LabelSprite position={[0, dy / 2 + 0.6, 0]} text={label} />
-    </group>
-  )
-}
-
-function AgentMesh({ getPos }: { getPos: () => Vec3 }) {
-  const ref = useRef<THREE.Mesh>(null!)
-  useFrame(() => {
-    const [x, y, z] = getPos()
-    ref.current.position.set(x, y, z)
-  })
-  return (
-    <mesh ref={ref} castShadow>
-      <sphereGeometry args={[0.35, 24, 24]} />
-      <meshStandardMaterial color="#4ade80" />
-    </mesh>
-  )
 }
 
 export default function BunkerPage() {
@@ -1015,8 +378,8 @@ export default function BunkerPage() {
 
         <div className="w-full h-[80vh] bg-black rounded-lg overflow-hidden">
           <Canvas shadows camera={{ position: [0, 12, 24], fov: 50 }}>
-		  	<CameraControls makeDefault />
-			<ambientLight intensity={0.6} />
+          	<CameraControls makeDefault />
+            <ambientLight intensity={0.6} />
             <directionalLight position={[10, 20, 10]} intensity={0.9} castShadow />
             <AnimateController />
